Move related event things in Event folder
