+++ src/Wallabag/CoreBundle/Event/Listener/LocaleListener.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Wallabag\CoreBundle\EventListener;
+namespace Wallabag\CoreBundle\Event\Listener;
 
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\HttpKernel\Event\GetResponseEvent;
